codexion/apps/middleware/requirements.txt
fastapi
uvicorn[standard]
bleak
cryptography
httpx
pydantic
python-dotenv
